pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "swreq"
version = "0.2.0"
description = "Two-stage LLM pipeline: C codebase → Low-Level Requirements → Jama-compliant HLRs via Ollama"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["requirements engineering", "LLM", "Jama", "MBSE", "software requirements", "Ollama"]

dependencies = [
    "pycparser>=2.21",
    "regex>=2023.6.3",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "streamlit>=1.34.0",
    "requests>=2.31.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",
    "python-docx>=1.1.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
swreq      = "main:main"
swreq-serve = "scripts.serve:main"

[tool.setuptools.packages.find]
where   = ["."]
include = ["src*", "api*"]

[tool.ruff]
line-length    = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]
